+++ utils/gob.go
@@ -1,40 +1,41 @@
 package utils
 
 import (
-	"io/ioutil"
 	"log"
 	"os"
-	"fmt"
+	"encoding/gob"
 )
 
-// NB: it is often faster to pass by value rather than difference in golang
-//FileToString gets the content of a file as a string
-func FileToString(filePath string) string {
-	dat, err := ioutil.ReadFile(filePath)
+// WriteGob writes the content of an interface variable to a file
+func WriteGob(filePath string, object interface{}) error {
+	file, err := os.Create(filePath)
+	if err!=nil {
+		log.Printf("Error trying to create \"%s\": %v", filePath, err)
+		return err
+	}
+	encoder := gob.NewEncoder(file)
+	encErr := encoder.Encode(object)
+	if encErr!=nil {
+		log.Printf("Error trying to encode \"%s\": %v", filePath, encErr)
+		return encErr
+	}
+	file.Close()
+	return nil
+}
+
+// ReadGob reads an interface variable from a file
+func ReadGob(filePath string, object interface{}) error {
+	file, err := os.Open(filePath)
 	if err != nil {
-		log.Printf("Error \"%v\" while trying to read \"%s\"\n", err, filePath)
+		log.Printf("Error trying to read %s: %v", filePath, err)
+		return err
 	}
-	fileString := string(dat)
-	return fileString
-}
-
-// ClearFolder clears a folder's content
-func ClearFolder(folderPath string) {
-	err := os.RemoveAll(folderPath)
-	if err!=nil{
-		fmt.Println(err)
+	decoder := gob.NewDecoder(file)
+	decErr := decoder.Decode(object)
+	if decErr != nil {
+		log.Printf("Error trying to decode %s, %v", filePath, decErr)
+		return decErr
 	}
-	err = os.MkdirAll(folderPath,0666)
-	if err!=nil{
-		fmt.Println(err)
-	}
-}
-
-// CheckPathExists returns true if a path exists, false otherwise
-func CheckPathExists(path string) bool {
-	_, err := os.Stat(path)
-	if os.IsNotExist(err) {
-		return false
-	}
-	return true
+	file.Close()
+	return nil
 }
